\name{hedo_biscuit}
\alias{hedo_biscuit}

\title{
Consumer data
}
\description{
hedo_biscuit data was collected from ratings of 294 consumers towards 8 products (cookies).
}
\usage{
hedo_biscuit
}
\format{
\describe{
  \item{\code{produit }}{
    factor : numbering the 8 products
  }
  \item{\code{consumers}}{
    294 consumers : ratings of consumers
  }
}}
\value{No return value, called for data loading}
\source{
  Francois Husson data sets
}

\references{
Pages et al. Sensory analysis comparison of eight biscuits
 by French and Pakistani panels, 2007, Journal of Sensory Studies 22(6), 665-686.7
}


\keyword{hedo_biscuit}

