\name{sens_biscuit}
\alias{sens_biscuit}

\title{
Sensory data
}
\description{
A sensory evaluation was organized in Pakistan to characterize
8 cookies products with a set of 23 quantitative sensory
attributes by 12 trained panelists at 2 tasting sessions.
}
\usage{
sens_biscuit
}
\format{
\describe{
  \item{\code{seance}}{
    factor : the tasting sessions 1 and 2
  }
  \item{\code{juge}}{
    factor : numbering the 12 assessors
  }
  \item{\code{produit}}{
     factor : numbering the 8 products
  }
  \item{\code{23 attributes}}{
    quantitative variables : characteristics of products
  }

}}
\value{No return value, called for data loading}
\source{
  Francois Husson data sets
}

\references{
Pages et al. Sensory analysis comparison of eight biscuits
 by French and Pakistani panels, 2007, Journal of Sensory Studies 22(6), 665-686.7
}


\keyword{sens_biscuit}

