\name{SensMapUI}
\alias{SensMapUI}


\title{Graphical User Interface for SensMap package}

\description{
launches a 'shiny' application that provides a Graphical User Interface (GUI) for the functions included in the SensMap package as well as a comparative tool of fit models performance using several criteria. Basic analysis such as characterization of products, panelists and sessions likewise consumers segmentation are also available. The application includes data import in different formats, plot displays and results in tables as well as downloading them. This application is very helpful for practitioners and researchers in sensory analysis field.
}
\usage{
SensMapUI()
}
\value{No return value, called for launching application}
\author{
Ibtihel Rebhi, Dhafer Malouche
}

\seealso{\link{PrefMap}, \link{SmoothMap}, \link{StabMap}}

\examples{
\dontrun{
library(SensMap)
SensMapUI()

}
}
