\name{SensMapUI}
\alias{SensMapUI}


\title{Graphical User Interface for SensMap package}

\description{
launches a 'shiny' application that provides a Graphical User Interface (GUI) for the functions included in the SensMap package as well as a comparative tool of fit models performance using several criteria. Basic analysis such as characterization of products, panelists and sessions likewise consumers segmentation are also available. This application is very helpful for researchers and practitioners. The application also includes data import in different formats, plot displays and results in tables as well as downloading them.
}
\usage{
SensMapUI()
}

\author{
Ibtihel Rebhi, Rihab Boubakri, Dhafer Malouche
}

\seealso{\link{PrefMap}, \link{SmoothMap}, \link{StabMap}}

\examples{

\dontrun{
library(SensMap)
SensMapUI()
}

}
