\name{SensMapUI}
\alias{SensMapUI}
\title{Graphical User interface for the SensMap package}
\usage{
SensMapUI()
}
\value{
Shiny application
}
\description{
launches a shiny application that provides the graphical
user interface (GUI) for the functions contained in the SensMap package. The application allows the importing of data in different formats and the presenting of results in tables and plots as well as downloading them.
}
\examples{
\dontrun{
library(SensMap)
SensMapUI()
}
}
\author{
Rebhi Ibtihel, Boubakri Rihab, Dhafer Malouche
}

